convert to size_t datatype

+++ src/Forest.h
@@ -99,8 +99,8 @@ public:
     }
     return result;
   }
-  std::vector<std::vector<double>> getSplitDecreases() {
-      std::vector<std::vector<double>> result;
+  std::vector<std::vector<size_t>> getSplitDecreases() {
+      std::vector<std::vector<size_t>> result;
       for (auto& tree : trees) {
           result.push_back(tree->getSplitDecreases());
       }
