+++ src/Tree.h
@@ -28,7 +28,7 @@ public:
 
   // Create from loaded forest
   Tree(std::vector<std::vector<size_t>>& child_nodeIDs, std::vector<size_t>& split_varIDs,
-      std::vector<double>& split_values, std::vector<double>& split_decreases, std::vector<double>& split_num_samples);
+      std::vector<double>& split_values, std::vector<size_t>& split_decreases, std::vector<double>& split_num_samples);
 
   virtual ~Tree() = default;
 
@@ -61,7 +61,7 @@ public:
   const std::vector<double>& getSplitValues() const {
     return split_values;
   }
-  const std::vector<double>& getSplitDecreases() const {
+  const std::vector<size_t>& getSplitDecreases() const {
       return split_decreases;
   }
   const std::vector<double>& getSplitNumSamples() const {
@@ -186,7 +186,7 @@ protected:
   std::vector<double> split_values;
 
   // TEST ZONE - split_decrease
-  std::vector<double> split_decreases;
+  std::vector<size_t> split_decreases;
   std::vector<double> split_num_samples;
 
   // Vector of left and right child node IDs, 0 for no child
